pyproject.toml
[tool.poetry]
name = "gps_synth"
version = "1.0.0"
description = "Python package to generate GPS synthetic data"
authors = ["Nikolay Kozlovskiy <nikolajkozlovskij73@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<=3.12"
geopandas = "0.14.0"
osmnx = "1.6.0"
scipy = "1.11.3"
pyarrow = "13.0.0"
PyYAML = "6.0.1"

[tool.poetry.group.vis]
optional = true

[tool.poetry.group.vis.dependencies]
jupyterlab = "3.6.3"
plotly = "5.16.1"
seaborn = "0.12.2"
matplotlib = "^3.5.2"
ipyleaflet = "0.17.3"
keplergl = "^0.3.2"
folium = "^0.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
